+++ meta/rescue.rb
@@ -9,6 +9,8 @@ Mutant::Meta::Example.add do
   mutation 'begin; rescue ExceptionA, self => error; true; end'
   mutation 'begin; rescue ExceptionA, ExceptionB => error; false; end'
   mutation 'begin; rescue ExceptionA, ExceptionB => error; nil; end'
+  mutation 'begin; true; end'
+
 end
 
 Mutant::Meta::Example.add do
@@ -19,6 +21,7 @@ Mutant::Meta::Example.add do
   mutation 'begin; rescue SomeException => error; false; end'
   mutation 'begin; rescue SomeException => error; nil; end'
   mutation 'begin; rescue self => error; true; end'
+  mutation 'begin; true; end'
 end
 
 Mutant::Meta::Example.add do
@@ -28,6 +31,7 @@ Mutant::Meta::Example.add do
   mutation 'begin; rescue => error; false; end'
   mutation 'begin; rescue => error; nil; end'
   mutation 'begin; rescue; true; end'
+  mutation 'begin; true; end'
 end
 
 Mutant::Meta::Example.add do
@@ -36,4 +40,5 @@ Mutant::Meta::Example.add do
   singleton_mutations
   mutation 'begin; rescue; false; end'
   mutation 'begin; rescue; nil; end'
+  mutation 'begin; true end'
 end