+++ Changelog.md
@@ -1,6 +1,7 @@
 # v0.6.1 2014-08-16
 
-* Add rescue else body promotion/concatenation mutation
+* Add rescue resbody body concat-promotion mutation
+* Add rescue else body concat-promotion mutation #245
 
 # v0.6.0 2014-08-11